Spilt webgpu ipc messages even more
