+++ components/webgpu/ipc_messages/mod.rs
@@ -1,0 +1,3 @@
+pub mod recv;
+pub mod send_dom;
+pub mod send_script;
