+++ components/webgpu/lib.rs
@@ -28,13 +28,13 @@ use webrender_traits::{
 };
 use wgc::id;
 
-mod dom_messages;
+mod ipc_messages;
 mod gpu_error;
 mod render_commands;
-mod script_messages;
-pub use dom_messages::*;
+pub use ipc_messages::send_dom::*;
+pub use ipc_messages::send_script::*;
+pub use ipc_messages::recv::*;
 pub use identity::*;
-pub use script_messages::*;
 pub use wgpu_thread::PRESENTATION_BUFFER_COUNT;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
